% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkCorNetwork.R
\name{benchmarkCorNetwork}
\alias{benchmarkCorNetwork}
\title{Benchmark Correlation-Score Network against a Reference interaction table}
\usage{
benchmarkCorNetwork(corNetwork, reference = "corum.txt", output = "pdf_csv")
}
\arguments{
\item{corNetwork}{Path to tab-separated .txt file or data.table with column headers uniprot_a uniprot_b cor_pearson and cor_spearman}

\item{reference}{Reference table, tab-separated text with headers uniprot_a and uniprot_b for true interactions.
Can also point to data.table with this information structure.}

\item{output}{Character string, wether .pdfs and .csvs shall be written to working folder or whether the output should go to
Rconsole. Options are "pdf_csv" or "Rconsole" with default "pdf_csv".}
}
\value{
data.frame with column headers as defined by arg protcolnames and cor_pearson and cor_spearman
}
\description{
Benchmark Correlation-Score Network against a Reference interaction table
}

