% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkNetwork.R
\name{benchmarkNetwork}
\alias{benchmarkNetwork}
\title{Benchmarks a Network against a Reference Network/Interaction set}
\usage{
benchmarkNetwork(Network, Reference = "intact.txt",
  names_net = c("UniprotA", "UniprotB"), scores_net = c("pW", "pNI",
  "pInt"), names_ref = c("uniprot_a", "uniprot_b", "miscore"),
  minscore_ref = 0.5, n_shuffle_iter = 10, output = "pdf_csv")
}
\arguments{
\item{Network}{data.table or path to tab-separated .txt file or data.table with columns uniprot_a uniprot_b and confidence score}

\item{Reference}{Reference table, data.table or path to tab-separated text with headers containing info uniprot_a and uniprot_b for true interactions.
Can also point to data.table with this information structure.}

\item{names_net}{Character vector specifiying the column names in Network containing the uniprot identifiers a and b and the confidence score.
Defaults to c("uniprot_a", "uniprot_b")}

\item{scores_net}{Character vector of score column names in the network data. Defaults to the Bioplex scores c("pW", "pNI", "pInt").}

\item{names_ref}{Character vector specifiying the column names in Network containing the uniprot identifiers a and b and the confidence score.
Defaults to c("uniprot_a", "uniprot_b" , "score")}

\item{minscore_ref}{minimum score threshold in the reference.}

\item{n_shuffle_iter}{Number of shufflings to generate the negative test set.}

\item{output}{Character string, wether .pdfs and .csvs shall be written to working folder or whether the output should go to
Rconsole. Options are "pdf_csv" or "Rconsole" with default "pdf_csv".}
}
\value{
data.frame with column headers as defined by arg protcolnames and cor_pearson and cor_spearman
}
\description{
Benchmarks a Network against a Reference Network/Interaction set
}

