% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSibTransCorr.R
\name{calculateSibTransCorr}
\alias{calculateSibTransCorr}
\title{Calculate sibling transition/fragment ion correlation in fragmention level
OpenSWATH/mProphet result table}
\usage{
calculateSibTransCorr(data.transition, plot = TRUE, PDF = FALSE)
}
\arguments{
\item{data.transition}{Data.table of disaggregated OpenSWATH data An object of type \code{traces.obj}.}

\item{plot}{logical TRUE or FALSE}

\item{PDF}{logical TRUE or FALSE}
}
\value{
data.transition Data.table of disaggregated OpenSWATH data including SibTransCorr column.
}
\description{
Calculate sibling transition/fragment ion correlation in fragmention level
OpenSWATH/mProphet result table
}

