% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conSecFilter.R
\name{conSecFilter}
\alias{conSecFilter}
\title{conSecFilter}
\usage{
conSecFilter(peptideprofiles, min_consecutive_ids = 3, remove_empty = TRUE)
}
\arguments{
\item{peptideprofiles}{data.frame with protein_id, peptide_id + fraction number columns, filled with intensities}

\item{min_consecutive_ids}{minimum required consecutive ids to keep quantifications}

\item{remove_empty}{whether rows that are empty after filtering shall be removed (i.e. peptide entries with less than min_consecutive_ids)}
}
\value{
filtered data.frame of same structure as input peptideprofiles table
}
\description{
conSecFilter
}

