% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenCorrMatrix.R
\name{flattenCorrMatrix}
\alias{flattenCorrMatrix}
\title{Flattens Correlation Matrix to table with columns row, column and cor}
\usage{
flattenCorrMatrix(cormat)
}
\arguments{
\item{cormat}{correlation matrix with row- and colnames as produced by cor() function}
}
\value{
data.frame with columns row, column and cor
}
\description{
Flattens Correlation Matrix to table with columns row, column and cor
}

