% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorNetwork.R
\name{generateCorNetwork}
\alias{generateCorNetwork}
\title{Generate Correlation-Score Network Table from SEC protein.traces}
\usage{
generateCorNetwork(protein.traces, cutoff = -1,
  protcolnames = c("uniprot_a", "uniprot_b"))
}
\arguments{
\item{cutoff}{Correlation score cutoff to apply before returning the output table}

\item{protcolnames}{Allows to set the column names of the protein-identifier containing columns.
Defaults to c("bait_id", "protein_id"), alternative, e.g.: protcolnames = c("uniprot_a", "uniprot_b").}

\item{rams}{protein.traces data.table with protein.traces, uniprot ids in column 1}
}
\value{
data.frame with column headers as defined in protcolnames (uniprot_a uniprot_b) and
         cor_pearson and cor_spearman.
}
\description{
Generate Correlation-Score Network Table from SEC protein.traces
}

