% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFFTs.R
\name{getFFTs}
\alias{getFFTs}
\title{Get fraction of false targets}
\usage{
getFFTs(folder = "pyprophet_stats/", pattern = "stats.csv")
}
\arguments{
\item{folder}{folder where the *full_stat.csv files are}

\item{pattern}{regex pattern for file grabbing}
}
\description{
Retrieve m/py-Prophet-model-estimated fraction of false targets (FFT) per run
}

