% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intactParser.R
\name{intactParser}
\alias{intactParser}
\title{Intact Parser}
\usage{
intactParser(intact, canonical = TRUE)
}
\arguments{
\item{intact}{intact data.table or path to .txt file}

\item{canonical}{whether uniprot entries shall be flattened to canonical, i.e. isoform flags (-x) removed}
}
\value{
data.table with columns uniprot_a uniprot_b and miscore
}
\description{
Parses intact.txt or data.table to 3-column format with clean uniprot_a uniprot_b and miscore columns
}

