% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairVenn.R
\name{pairVenn}
\alias{pairVenn}
\title{pairVenn
 @description Draws a pairwise Venn Diagram of two character vectors with overlap area
 corresponding to overlap size.
 @param a Character vector a
 @param b Character vector b
 @param output character selection either "R" or "pdf"
 @import VennDiagram
 @import gridExtra
 @export}
\usage{
pairVenn(a, b, output = "R")
}
\description{
pairVenn
 @description Draws a pairwise Venn Diagram of two character vectors with overlap area
 corresponding to overlap size.
 @param a Character vector a
 @param b Character vector b
 @param output character selection either "R" or "pdf"
 @import VennDiagram
 @import gridExtra
 @export
}

