% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFilterImpact.R
\name{plotFilterImpact}
\alias{plotFilterImpact}
\title{Plot Peptide Elution Profiles per Protein without and with Filters Applied}
\usage{
plotFilterImpact(traces, traces.sf = NULL, sf_setting = 3,
  traces.cf = NULL, cf_setting = 0.2)
}
\arguments{
\item{traces}{A traces object containing $peptide.profiles data.table with protein_id and peptide_id col}

\item{traces.sf}{A traces object containing stretch-filtered $peptide.profiles as generated by filterConsecutiveIdStretches().
data.table with protein_id and peptide_id col. If not provided, it is generated.}

\item{sf_setting}{min_stretch_length setting that is used to generate the stretch-filtered peptide traces data if not provided.
Defaults to 3.}

\item{traces.cf}{A traces object containing correlation-filtered $peptide.profiles as generated by filterByPeptideCorrelation()
data.table with protein_id and peptide_id col. If not provided, it is generated.}

\item{cf_setting}{average_corr_cutoff setting that is used to generate the correlation-filtered peptide traces data if not provided.
Defaults to 0.2.}
}
\value{
A data.table with the number of proteins and peptides remaining after filtering.
        Generates a FilterImpact.pdf in the working folder.
}
\description{
Native elution profiling data can be noisy. To suppress noise and select for the high quality signals, filters can be applied.
       This function wraps the filtering functions filterConsecutiveIdStretches() and filterByPeptideCorrelation() and visualizes their impact
       on protein elution profiles by combined plots of all peptides of a protein over the elution fraction number.
}
\examples{
# NOT RUN:
# filterImpactNumbers <- plotFilterImpact(peptide.traces)
}

