% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPANTHERresult.R
\name{plotPANTHERresult}
\alias{plotPANTHERresult}
\title{Plot PantherDB enrichment/overrepresentation test result}
\usage{
plotPANTHERresult(PANTHERresult = "panther_gocc.txt", p_cutoff = 0.05)
}
\arguments{
\item{PANTHERresult}{Panther result table path to .txt or data.table}

\item{p_cutoff}{Significance Cutoff for labelling the found terms in the plots}
}
\value{
A list of 1) the reformatted data.table, 2) scatter plot 3) barplot
}
\description{
Plot PantherDB enrichment/overrepresentation test result
}
\examples{
plotPANTHERresult("panther_gocc.txt", p_cutoff = 0.05)
}

