% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeptideProfiles.R
\name{plotPeptideProfiles}
\alias{plotPeptideProfiles}
\title{plotPeptideProfiles}
\usage{
plotPeptideProfiles(Traces, select_proteins = "all", output = "pdf")
}
\arguments{
\item{Traces}{list with $peptide.traces data tabl with protein_id, peptide_id and ordered quant values in the remaining columns}

\item{select_proteins}{character vector of protein_ids}
}
\description{
plot all peptide SEC chromatograms per protein
}

