% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPepSignals.R
\name{sumPepSignals}
\alias{sumPepSignals}
\title{sumPepSignals}
\usage{
sumPepSignals(widepepdata)
}
\arguments{
\item{widepepdata}{wide peptide level quant data.table with label columns protein_id and peptide_id}
}
\value{
Wide protein level table with peptide signals summed per protein_id
}
\description{
sumPepSignals
}

