% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairVenn.R
\name{pairVenn}
\alias{pairVenn}
\title{pairVenn}
\usage{
pairVenn(a, b, output = "R")
}
\arguments{
\item{a}{Character vector a}

\item{b}{Character vector b}

\item{output}{character selection either "R" or "pdf"}
}
\value{
paiwise Venn Diagram plot with areas corresponding to numbers
}
\description{
Draws a pairwise Venn Diagram of two character vectors with overlap area corresponding to overlap size.
}

